backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
librosa==0.10.1
numpy==1.26.3
scipy==1.12.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
boto3==1.34.34
yt-dlp==2024.1.7
ffmpeg-python==0.2.0
pytest==7.4.4
pytest-asyncio==0.23.3
